libgo: 1.25 runtime reconciliation + AddCleanup (WIP: stdlib tail remains)
Runtime package now compiles and runtime.lo builds. Reconciled ~105 runtime
errors from the gc1.24->1.25 merge (reject Green Tea GC, AddCleanup specials/
checkfinalizer, pclntab refs, swiss; keep gccgo runtime): allArenas->heapArenas,
sysAllocOS/sysAlloc/sysReserveAligned vmaName params, size-class dedup vs
internal/runtime/gc, cleanup-special removal across mgc/mgcmark/mgcsweep/mheap/
mprof/proc/mcheckmark, ncpu->numCPUStartup, alg readUnaligned restore,
set_vma_name_linux syscall fix, m.allpSnapshot, gFree flat, gcount, misc fields.
synctest: internal/synctest bodyless pulls + runtime pushes (mangled
internal_1synctest.*), synctestmain named func (no closure escape), panic.go
cgroup_throw target mangled internal_1runtime_1cgroup.throw.

Feature: implement runtime.AddCleanup/Cleanup in gccgo (mcleanup_gccgo.go) on
top of SetFinalizer/removefinalizer (gccgo has no gc cleanup machinery); unblocks
os/net/crypto-tls/fips140cache/unique/testing-synctest. Deleted orphaned
unique/canonmap.go (gccgo unique uses sync.Map).

Stdlib fixes: at_sysnum_linux fchmodat/fchownat/linkat/symlinkat Trap consts,
time.runtimeIsBubbled bodyless, WCLONE for non-loong64 linux, xor_asm && gc,
internal/poll SendFile src uintptr + splice restore (finalizer-based).

REMAINING (stdlib tail): reflect *abi.Type vs *rtype bridge (json/xml/asn1/sql/
text-template/x-tools), net sendFile redefinition + SendFile int->uintptr
cascade, runtime/trace flightrecorder/subscribe linknames, frontend ICE
parse.cc:7695 on x/tools typeutil generics.

+++ libgo/go/internal/poll/sendfile_glibc.go
@@ -13,7 +13,7 @@ import "syscall"
 const maxSendfileSize int = 4 << 20
 
 // SendFile wraps the sendfile system call.
-func SendFile(dstFD *FD, src int, remain int64) (int64, error, bool) {
+func SendFile(dstFD *FD, src uintptr, remain int64) (int64, error, bool) {
 	if err := dstFD.writeLock(); err != nil {
 		return 0, err, false
 	}
@@ -33,7 +33,7 @@ func SendFile(dstFD *FD, src int, remain int64) (int64, error, bool) {
 		if int64(n) > remain {
 			n = int(remain)
 		}
-		n, err1 := syscall.Sendfile(dst, src, nil, n)
+		n, err1 := syscall.Sendfile(dst, int(src), nil, n)
 		if n > 0 {
 			written += int64(n)
 			remain -= int64(n)
